Code fail fixes

+++ main.h
@@ -22,7 +22,7 @@ int Putchar(char ch);
 int Puts(char *string);
 int _printf(const char *format, ...);
 char *ConvertBase(uint64_t number, uint8_t radix);
-int DecimalToBinary(uint64_t decimalNum);
+int DecimalToBinary(uint32_t decimalNum);
 int PutsCustom(char *string);
 int PrintfFlags(char flag, va_list arguments);
 int PutsReverse(char *string);
